Fixed Mage's Staff when imbuing unenchanted Magic Missile but Staff is enchanted.

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Mob.java
@@ -584,7 +584,7 @@ public abstract class Mob extends Char {
 			alerted = true;
 		}
 		if (Dungeon.isChallenged(Challenges.EVIL_GAS)) {
-			GameScene.add(Blob.seed(this.pos, 50*(dmg/HT), EvilGas.class).setStrength(1+Dungeon.depth/4));
+			GameScene.add(Blob.seed(this.pos, (int) (50*((float)dmg/(float)HT)), EvilGas.class).setStrength(1+Dungeon.depth/4));
 		}
 		super.damage( dmg, src );
 	}
